Add a projects folder that can safely be ignored in git
Cloning projects into the vagrant repo to work with inside and out of the VM makes it hard to make changes to the vagrant repo itself.  This creates a folder that can still be accessed inside and out of the VM, and not interfear with the changes going on to the vagrant file.